Load module before doing most of the rest of the saml config
